cafe_bot/requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
apscheduler==3.8.1  
regex==2022.10.31  
aiofiles==24.1.0
